+++ app/src/main/java/ru/grebnev/repairestimate/project/dialogs/EditingProjectDialog.java
@@ -26,7 +26,7 @@ public class EditingProjectDialog extends DialogFragment {
 
         args.putLong("date", project.getDateProject());
         args.putString("name", project.getNameProject());
-        args.putInt("sum", project.getSumProject());
+        args.putFloat("sum", project.getSumProject());
 
         editingProjectDialog.setArguments(args);
         return editingProjectDialog;
